+++ src/scripts/CreateLine.tsx
@@ -127,24 +127,3 @@ export function CreateHtmlLine(type: string, dialogue: string, speaker?: string,
 }
 
 
-export function CreateEmLine(position: string, icon: JSX.Element): JSX.Element {
-
-    let speaker = fetchSpritePos(position)
-    if (speaker === 'null') {
-        throw new Error("LADR: Speaker value returned null.")
-    }
-
-    return (
-        <div className="flex p-2 rounded noto-serif-kr">
-            <div id="speaker" className="flex justify-end flex-shrink-0 w-40 mr-2">
-                <p className="font-semibold text-gray-600">{speaker}:</p>
-            </div>
-            <div id="dialogue" className="text-gray-200">
-                {icon}
-            </div>
-        </div>
-
-    )
-
-
-}